pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "donna-cli"
version = "0.1.0"
description = "Digital Operative for Non-Negotiable Automation — a CLI-resident multi-agent framework."
readme = "README.md"
license-expression = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Trishul" },
]
keywords = ["ai", "cli", "agent", "llm", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "typer[all]>=0.9",
    "prompt-toolkit>=3.0",
    "rich>=13.0",
    "httpx>=0.27",
    "openai>=1.0",
    "pyperclip>=1.8",
    "chromadb>=0.4",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
donna = "donna.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["donna", "config"]

